Add files via upload
This file reads segmented contour image files on a day-by-day basis. It then extracts all the connected components and saves each connected component in a separate folder called connected component. It then reads these connected components and classifies them as an interest image or a non-interest image. Once this is done, it iterates over the interest image files, and gets relevant information (Day, Center number, connected component number, x and y coordinate of the center) and saves it in a text file. Once that is done it converts the text file into a csv file with correct column headers.