Add files via upload
This code takes in the csv file that contains the extracted day-center information. It then iterates through all the points in the center and gives them a class label from a different file called 15x15_area_f.csv (it contains the class value of each and every center that we are trying to classify). It then saves a new csv file that containing information regarding all interest points, their Day, class name, width and height (of the cluster they are associated with), etc.