Add files via upload
This file creates a l dataframe containing all the possible days, their month and year values and extracts those days for our testing and training dataset (testing month 7 and year 4 and 8, training month 7 and all remaining years). It then iterates through the file containing the class values and the center information. It creates a new dataset, which contains the x, and y coordinate and the pressure data of the two previous days within a specified region as well as the class value. This region will be a rectangle have the shape of 30x30 pixels. the center of this rectangle will be the x,y coordinates of the point itself. So basically we are extracting the pressure data from a separate folder, and taking the 30x30 pixels surrounding the x,y coordinates on the two previous days. This is done so that we understand what is going on in the two previous days and can predict the class value of the current day.