Add files via upload
This file creates and deploys a CNN model. It trains the data based on a training set, and tests it based on the testing set. It also visualizes the results by saving the predicted image, the true image and the correctly predicted image in separate files. It also creates a csv file containing the accuracy, as well as other relevant information for analysis of the results of the model.